+++ impls/agents/__init__.py
@@ -1,7 +1,6 @@
 from agents.crl import CRLAgent
 from agents.gcbc import GCBCAgent
 from agents.gciql import GCIQLAgent
-from agents.gciql_chunk import GCIQLChunkAgent
 from agents.gcivl import GCIVLAgent
 from agents.hiql import HIQLAgent
 from agents.qrl import QRLAgent
@@ -11,7 +10,6 @@ agents = dict(
     crl=CRLAgent,
     gcbc=GCBCAgent,
     gciql=GCIQLAgent,
-    gciql_chunk=GCIQLChunkAgent,
     gcivl=GCIVLAgent,
     hiql=HIQLAgent,
     qrl=QRLAgent,